--- Matthijs/conveyorbelt.cpp
@@ -1,11 +1,0 @@
-#include "conveyorbelt.h"
-#include "block.h"
-
-conveyorBelt::conveyorBelt()
-{
-public:
-    conveyorBelt();
-
-    void setOrder(block order[8]);
-
-}
