--- Matthijs/isensor.cpp
@@ -1,12 +1,0 @@
-#include "isensor.h"
-
-ISensor::ISensor()
-{
-public:
-    ISensor();
-
-    virtual bool getSensorStatus() = 0;
-
-private:
-    bool waarde;
-}
