--- Matthijs/metalsensor.cpp
@@ -1,13 +1,0 @@
-#include "metalsensor.h"
-
-metalSensor::metalSensor()
-{
-public:
-    metalSensor();
-
-    bool getSensorStatus();
-
-private:
-    bool status;
-
-}
