--- Matthijs/motor.cpp
@@ -1,13 +1,0 @@
-#include "motor.h"
-
-motor::motor()
-{
-public:
-    motor();
-
-    void setMotor();
-    bool getMotorStatus();
-
-private:
-    bool motorOn;
-}
