--- Matthijs/pushrod.cpp
@@ -1,13 +1,0 @@
-#include "pushrod.h"
-
-pushRod::pushRod()
-{
-public:
-    pushRod();
-
-    void setPushRod(bool _status);
-    bool getPushRodStatus();
-
-private:
-    bool pushRodActive;
-}
