+++ slotsandsignals/.gitignore
