+++ slotsandsignals/block.cpp
