+++ slotsandsignals/block.h
@@ -6,6 +6,11 @@ class block
 {
 public:
     block();
+    block(int _height, bool _metal);
+
+private:
+    int height;
+    bool metal;
 };
 
 #endif // BLOCK_H
