+++ slotsandsignals/conveyorbelt.cpp
@@ -1,0 +1,7 @@
+#include "conveyorbelt.h"
+
+conveyorBelt::conveyorBelt()
+{
+a
+
+}
