+++ slotsandsignals/conveyorbelt.h
@@ -1,11 +1,14 @@
 #ifndef CONVEYORBELT_H
 #define CONVEYORBELT_H
 
+#include "block.h"
 
 class conveyorBelt
 {
 public:
     conveyorBelt();
+
+    void setOrder(block order[8]);
 };
 
 #endif // CONVEYORBELT_H
