+++ slotsandsignals/distancesensor.cpp
