+++ slotsandsignals/distancesensor.h
