+++ slotsandsignals/isensor.cpp
@@ -1,0 +1,6 @@
+#include "isensor.h"
+
+ISensor::ISensor()
+{
+
+}
