+++ slotsandsignals/isensor.h
@@ -6,6 +6,11 @@ class ISensor
 {
 public:
     ISensor();
+
+    virtual bool getSensorStatus() = 0;
+
+private:
+    bool waarde;
 };
 
 #endif // ISENSOR_H
