+++ slotsandsignals/lowDistancesensor.cpp
@@ -1,0 +1,7 @@
+#include "lowDistancesensor.h"
+#include "isensor.h"
+
+lowDistanceSensor::lowDistanceSensor()
+{
+
+}
