+++ slotsandsignals/mainwindow.h
