+++ slotsandsignals/mainwindow.ui
