+++ slotsandsignals/metalsensor.cpp
@@ -1,0 +1,7 @@
+#include "metalsensor.h"
+
+metalSensor::metalSensor()
+{
+
+
+}
