+++ slotsandsignals/metalsensor.h
@@ -6,6 +6,11 @@ class metalSensor
 {
 public:
     metalSensor();
+
+    bool getSensorStatus();
+
+private:
+    bool status;
 };
 
 #endif // METALSENSOR_H
