+++ slotsandsignals/motor.cpp
@@ -1,0 +1,6 @@
+#include "motor.h"
+
+motor::motor()
+{
+
+}
