+++ slotsandsignals/motor.h
@@ -6,6 +6,12 @@ class motor
 {
 public:
     motor();
+
+    void setMotor();
+    bool getMotorStatus();
+
+private:
+    bool motorOn;
 };
 
 #endif // MOTOR_H
