+++ slotsandsignals/pushrod.h
@@ -6,6 +6,12 @@ class pushRod
 {
 public:
     pushRod();
+
+    void setPushRod(bool _status);
+    bool getPushRodStatus();
+
+private:
+    bool pushRodActive;
 };
 
 #endif // PUSHROD_H
